requirements.txt
psycopg2-binary
pandas
pyyaml
python-dotenv
numpy
